Create 31_next_permutation.cpp
Next Permutation is the requires the user to print the next lexicographical permutation o[f numbers in the given array. They key to solving this problem is identifying the breakpoint which separates the array into the prefix which remains same between two consecutive permutations and the suffix which changes between two permutations.